% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_rtest_mul.R
\name{ff_rtest_mul_basic}
\alias{ff_rtest_mul_basic}
\title{Multiply together two numbers}
\usage{
ff_rtest_mul_basic(ar_fl_x, ar_fl_y)
}
\arguments{
\item{ar_fl_x}{Scalar or vector of numbers, same length as \code{ar_fl_y}}

\item{ar_fl_y}{Scalar or vector of numbers, same length as \code{ar_fl_x}}
}
\value{
The product of \code{ar_fl_x} and \code{ar_fl_y} inline equation \eqn{x \cdot y}.
}
\description{
This function provides an example with Roxygen2 markups.
We can internally reference the sister function \code{\link[=ff_rtest_add_basic]{ff_rtest_add_basic()}}. Display
equation is shown with \deqn{x \cdot y}
}
\examples{
ff_rtest_mul_basic(1, 1)
ff_rtest_mul_basic(c(10,10), c(1,3))
}
\references{
\url{https://cran.r-project.org/web/packages/roxygen2/vignettes/roxygen2.html}
}
\seealso{
\code{\link[=ff_rtest_add_basic]{ff_rtest_add_basic()}} for addition.
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
